split container test files
