+++ tests/playwright/sanity/modules/container/container-3.test.ts-snapshots/container-background-linux.jpeg
