+++ tests/playwright/sanity/modules/container/container-3.test.ts-snapshots/container-ltr-centered-linux.jpeg
